Crate google_plus1

This documentation was generated from plus crate version 0.1.8+20150303, where 20150303 is the exact revision of the plus:v1 schema built by the mako code generator v0.1.8.

Everything else about the plus v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

activities

get, list and search

comments

get and list

moments

insert, list and remove

people

get, list, list by activity and search

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.people().search(...).doit()
let r = hub.people().list(...).doit()
let r = hub.people().list_by_activity(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-plus1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_plus1 as plus1;
use plus1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plus1::Plus;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Plus::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.people().list("userId", "collection")
             .page_token("et")
             .order_by("eirmod")
             .max_results(43)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Acl
There is no detailed description.

Activity
There is no detailed description.

ActivityActor
The person who performed this activity.

ActivityActorImage
The image representation of the actor.

ActivityActorName
An object representation of the individual components of name.

ActivityFeed
There is no detailed description.

ActivityGetCall
Get an activity.

ActivityListCall
List all of the activities in the specified collection for a particular user.

ActivityMethods
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Plus hub.

ActivityObject
The object of this activity.

ActivityObjectActor
If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor.

ActivityObjectActorImage
The image representation of the original actor.

ActivityObjectAttachments
The media objects attached to this activity.

ActivityObjectAttachmentsEmbed
If the attachment is a video, the embeddable link.

ActivityObjectAttachmentsFullImage
The full image URL for photo attachments.

ActivityObjectAttachmentsImage
The preview image for photos or videos.

ActivityObjectAttachmentsThumbnails
If the attachment is an album, this property is a list of potential additional thumbnails from the album.

ActivityObjectAttachmentsThumbnailsImage
Image resource.

ActivityObjectPlusoners
People who +1'd this activity.

ActivityObjectReplies
Comments in reply to this activity.

ActivityObjectResharers
People who reshared this activity.

ActivityProvider
The service provider that initially published this activity.

ActivitySearchCall
Search public activities.

Comment
There is no detailed description.

CommentActor
The person who posted this comment.

CommentActorImage
The image representation of this actor.

CommentFeed
There is no detailed description.

CommentGetCall
Get a comment.

CommentInReplyTo
The activity this comment replied to.

CommentListCall
List all of the comments for an activity.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Plus hub.

CommentObject
The object of this comment.

CommentPlusoners
People who +1'd this comment.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ItemScope
There is no detailed description.

MethodInfo
Contains information about an API request.

Moment
There is no detailed description.

MomentInsertCall
Record a moment representing a user's action such as making a purchase or commenting on a blog.

MomentListCall
List all of the moments for a particular user.

MomentMethods
A builder providing access to all methods supported on moment resources. It is not used directly, but through the Plus hub.

MomentRemoveCall
Delete a moment.

MomentsFeed
There is no detailed description.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

PeopleFeed
There is no detailed description.

PeopleGetCall
Get a person's profile. If your app uses scope https://www.googleapis.com/auth/plus.login, this method is guaranteed to return ageRange and language.

PeopleListByActivityCall
List all of the people in the specified collection for a particular activity.

PeopleListCall
List all of the people in the specified collection.

PeopleMethods
A builder providing access to all methods supported on people resources. It is not used directly, but through the Plus hub.

PeopleSearchCall
Search all public profiles.

Person
There is no detailed description.

PersonAgeRange
The age range of the person. Valid ranges are 17 or younger, 18 to 20, and 21 or older. Age is determined from the user's birthday using Western age reckoning.

PersonCover
The cover photo content.

PersonCoverCoverInfo
Extra information about the cover photo.

PersonCoverCoverPhoto
The person's primary cover image.

PersonEmails
A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile. The plus.profile.emails.read scope is needed to retrieve these email addresses, or the email scope can be used to retrieve just the Google account email address.

PersonImage
The representation of the person's profile photo.

PersonName
An object representation of the individual components of a person's name.

PersonOrganizations
A list of current or past organizations with which this person is associated.

PersonPlacesLived
A list of places where this person has lived.

PersonUrls
A list of URLs for this person.

Place
There is no detailed description.

PlaceAddress
The physical address of the place.

PlacePosition
The position of the place.

Plus
Central instance to access all Plus related resource activities

PlusAclentryResource
There is no detailed description.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.