Module google_plus1::api

Structs

Acl
There is no detailed description.

Activity
There is no detailed description.

ActivityActor
The person who performed this activity.

ActivityActorClientSpecificActorInfo
Actor info specific to particular clients.

ActivityActorClientSpecificActorInfoYoutubeActorInfo
Actor info specific to YouTube clients.

ActivityActorImage
The image representation of the actor.

ActivityActorName
An object representation of the individual components of name.

ActivityActorVerification
Verification status of actor.

ActivityFeed
There is no detailed description.

ActivityGetCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

ActivityListCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

ActivityMethods
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Plus hub.

ActivityObject
The object of this activity.

ActivityObjectActor
If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor.

ActivityObjectActorClientSpecificActorInfo
Actor info specific to particular clients.

ActivityObjectActorClientSpecificActorInfoYoutubeActorInfo
Actor info specific to YouTube clients.

ActivityObjectActorImage
The image representation of the original actor.

ActivityObjectActorVerification
Verification status of actor.

ActivityObjectAttachments
The media objects attached to this activity.

ActivityObjectAttachmentsEmbed
If the attachment is a video, the embeddable link.

ActivityObjectAttachmentsFullImage
The full image URL for photo attachments.

ActivityObjectAttachmentsImage
The preview image for photos or videos.

ActivityObjectAttachmentsThumbnails
If the attachment is an album, this property is a list of potential additional thumbnails from the album.

ActivityObjectAttachmentsThumbnailsImage
Image resource.

ActivityObjectPlusoners
People who +1'd this activity.

ActivityObjectReplies
Comments in reply to this activity.

ActivityObjectResharers
People who reshared this activity.

ActivityProvider
The service provider that initially published this activity.

ActivitySearchCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

Comment
There is no detailed description.

CommentActor
The person who posted this comment.

CommentActorClientSpecificActorInfo
Actor info specific to particular clients.

CommentActorClientSpecificActorInfoYoutubeActorInfo
Actor info specific to YouTube clients.

CommentActorImage
The image representation of this actor.

CommentActorVerification
Verification status of actor.

CommentFeed
There is no detailed description.

CommentGetCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

CommentInReplyTo
The activity this comment replied to.

CommentListCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Plus hub.

CommentObject
The object of this comment.

CommentPlusoners
People who +1'd this comment.

PeopleFeed
There is no detailed description.

PeopleGetCall
Get a person's profile. If your app uses scope https://www.googleapis.com/auth/plus.login, this method is guaranteed to return ageRange and language.

PeopleListByActivityCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

PeopleListCall
List all of the people in the specified collection.

PeopleMethods
A builder providing access to all methods supported on people resources. It is not used directly, but through the Plus hub.

PeopleSearchCall
Shut down. See https://developers.google.com/+/api-shutdown for more details.

Person
There is no detailed description.

PersonAgeRange
The age range of the person. Valid ranges are 17 or younger, 18 to 20, and 21 or older. Age is determined from the user's birthday using Western age reckoning.

PersonCover
The cover photo content.

PersonCoverCoverInfo
Extra information about the cover photo.

PersonCoverCoverPhoto
The person's primary cover image.

PersonEmails
A list of email addresses that this person has, including their Google account email address, and the public verified email addresses on their Google+ profile. The plus.profile.emails.read scope is needed to retrieve these email addresses, or the email scope can be used to retrieve just the Google account email address.

PersonImage
The representation of the person's profile photo.

PersonName
An object representation of the individual components of a person's name.

PersonOrganizations
A list of current or past organizations with which this person is associated.

PersonPlacesLived
A list of places where this person has lived.

PersonUrls
A list of URLs for this person.

Place
There is no detailed description.

PlaceAddress
The physical address of the place.

PlacePosition
The position of the place.

Plus
Central instance to access all Plus related resource activities

PlusAclentryResource
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.