Struct google_plus1::ActivityMethods

pub struct ActivityMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on activity resources. It is not used directly, but through the Plus hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_plus1 as plus1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plus1::Plus;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Plus::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `list(...)` and `search(...)`
// to build up your call.
let rb = hub.activities();

Methods

impl<'a, C, A> ActivityMethods<'a, C, A>

fn search(&self, query: &str) -> ActivitySearchCall<'a, C, A>

Create a builder to help you perform the following task:

Search public activities.

Arguments

query - Full-text search query string.

fn get(&self, activity_id: &str) -> ActivityGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an activity.

Arguments

activityId - The ID of the activity to get.

fn list(&self, user_id: &str, collection: &str) -> ActivityListCall<'a, C, A>

Create a builder to help you perform the following task:

List all of the activities in the specified collection for a particular user.

Arguments

userId - The ID of the user to get activities for. The special value "me" can be used to indicate the authenticated user.

collection - The collection of activities to list.