Struct google_plus1::ActivityObject

pub struct ActivityObject {
    pub content: Option<String>,
    pub attachments: Option<Vec<ActivityObjectAttachments>>,
    pub original_content: Option<String>,
    pub plusoners: Option<ActivityObjectPlusoners>,
    pub actor: Option<ActivityObjectActor>,
    pub resharers: Option<ActivityObjectResharers>,
    pub url: Option<String>,
    pub replies: Option<ActivityObjectReplies>,
    pub id: Option<String>,
    pub object_type: Option<String>,
}

The object of this activity.

This type is not used in any activity, and only used as part of another schema.

Fields

content

The HTML-formatted content, which is suitable for display.

attachments

The media objects attached to this activity.

original_content

The content (text) as provided by the author, which is stored without any HTML formatting. When creating or updating an activity, this value must be supplied as plain text in the request.

plusoners

People who +1'd this activity.

actor

If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor.

resharers

People who reshared this activity.

url

The URL that points to the linked resource.

replies

Comments in reply to this activity.

id

The ID of the object. When resharing an activity, this is the ID of the activity that is being reshared.

object_type

The type of the object. Possible values include, but are not limited to, the following values: - "note" - Textual content. - "activity" - A Google+ activity.