Struct google_plus1::ItemScope

pub struct ItemScope {
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub text: Option<String>,
    pub image: Option<String>,
    pub review_rating: Option<Option<Box<ItemScope>>>,
    pub birth_date: Option<String>,
    pub date_published: Option<String>,
    pub address_locality: Option<String>,
    pub duration: Option<String>,
    pub additional_name: Option<Vec<String>>,
    pub worst_rating: Option<String>,
    pub contributor: Option<Vec<ItemScope>>,
    pub thumbnail_url: Option<String>,
    pub id: Option<String>,
    pub post_office_box_number: Option<String>,
    pub attendees: Option<Vec<ItemScope>>,
    pub author: Option<Vec<ItemScope>>,
    pub associated_media: Option<Vec<ItemScope>>,
    pub best_rating: Option<String>,
    pub address_country: Option<i64>,
    pub width: Option<String>,
    pub street_address: Option<String>,
    pub location: Option<Option<Box<ItemScope>>>,
    pub latitude: Option<f64>,
    pub by_artist: Option<Option<Box<ItemScope>>>,
    pub type_: Option<String>,
    pub date_modified: Option<String>,
    pub content_size: Option<String>,
    pub content_url: Option<String>,
    pub family_name: Option<String>,
    pub description: Option<String>,
    pub part_of_tv_series: Option<Option<Box<ItemScope>>>,
    pub about: Option<Option<Box<ItemScope>>>,
    pub date_created: Option<String>,
    pub postal_code: Option<String>,
    pub attendee_count: Option<i32>,
    pub in_album: Option<Option<Box<ItemScope>>>,
    pub address: Option<Option<Box<ItemScope>>>,
    pub height: Option<String>,
    pub geo: Option<Option<Box<ItemScope>>>,
    pub embed_url: Option<String>,
    pub ticker_symbol: Option<String>,
    pub name: Option<String>,
    pub kind: Option<String>,
    pub audio: Option<Option<Box<ItemScope>>>,
    pub player_type: Option<String>,
    pub performers: Option<Vec<ItemScope>>,
    pub url: Option<String>,
    pub gender: Option<String>,
    pub longitude: Option<f64>,
    pub thumbnail: Option<Option<Box<ItemScope>>>,
    pub address_region: Option<String>,
    pub caption: Option<String>,
    pub rating_value: Option<String>,
    pub given_name: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

start_date

The start date and time of the event (in ISO 8601 date format).

end_date

The end date and time of the event (in ISO 8601 date format).

text

The text that is the result of the app activity. For example, if a user leaves a review of a restaurant, this might be the text of the review.

image

A URL to the image that represents this result. For example, if a user writes a review of a restaurant and attaches a photo of their meal, you might use that photo as the result.image.

review_rating

Review rating.

birth_date

Date of birth.

date_published

The initial date that the result was published. For example, a user writes a comment on a blog, which has a result.dateCreated of when they submit it. If the blog users comment moderation, the result.datePublished value would match the date when the owner approved the message.

address_locality

Address locality.

duration

The duration of the item (movie, audio recording, event, etc.) in ISO 8601 date format.

additional_name

An additional name for a Person, can be used for a middle name.

worst_rating

Worst possible rating value that a result might obtain. This property defines the lower bound for the ratingValue.

contributor

A list of contributors to this result.

thumbnail_url

A URL to a thumbnail image that represents this result.

id

An identifier for the object. Your app can choose how to identify objects. The object.id is required if you are writing an action that does not have a corresponding web page or object.url property.

post_office_box_number

Post office box number.

attendees

A person attending the event.

author

The person or persons who created this result. In the example of restaurant reviews, this might be the reviewer's name.

associated_media

The encoding.

best_rating

Best possible rating value that a result might obtain. This property defines the upper bound for the ratingValue. For example, you might have a 5 star rating scale, you would provide 5 as the value for this property.

address_country

Address country.

width

The width of the media object.

street_address

Street address.

location

The location of the event or organization.

latitude

Latitude.

by_artist

From http://schema.org/MusicRecording, the artist that performed this recording.

type_

The schema.org URL that best describes the referenced object and matches the type of moment.

date_modified

The date the result was last modified such as the date that a review was last edited.

content_size

File size in (mega/kilo) bytes.

content_url

Actual bytes of the media object, for example the image file or video file.

family_name

Family name. This property can be used with givenName instead of the name property.

description

The string that describes the content of the result.

part_of_tv_series

Property of http://schema.org/TVEpisode indicating which series the episode belongs to.

about

The subject matter of the content.

date_created

The date the result was created such as the date that a review was first created.

postal_code

Postal code.

attendee_count

Number of attendees.

in_album

From http://schema.org/MusicRecording, which album a song is in.

address

Postal address.

height

The height of the media object.

geo

Geo coordinates.

embed_url

A URL pointing to a player for a specific video. In general, this is the information in the src element of an embed tag and should not be the same as the content of the loc tag.

ticker_symbol

The exchange traded instrument associated with a Corporation object. The tickerSymbol is expressed as an exchange and an instrument name separated by a space character. For the exchange component of the tickerSymbol attribute, we recommend using the controlled vocabulary of Market Identifier Codes (MIC) specified in ISO15022.

name

The name of the result. In the example of a restaurant review, this might be the summary the user gave their review such as "Great ambiance, but overpriced."

kind

Identifies this resource as an itemScope.

audio

From http://schema.org/MusicRecording, the audio file.

player_type

Player type that is required. For example: Flash or Silverlight.

performers

The main performer or performers of the event-for example, a presenter, musician, or actor.

url

The URL that points to the result object. For example, a permalink directly to a restaurant reviewer's comment.

gender

Gender of the person.

longitude

Longitude.

thumbnail

Thumbnail image for an image or video.

address_region

Address region.

caption

The caption for this object.

rating_value

Rating value.

given_name

Given name. This property can be used with familyName instead of the name property.