Struct google_plus1::PeopleMethods

pub struct PeopleMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on people resources. It is not used directly, but through the Plus hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_plus1 as plus1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plus1::Plus;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Plus::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `list(...)`, `list_by_activity(...)` and `search(...)`
// to build up your call.
let rb = hub.people();

Methods

impl<'a, C, A> PeopleMethods<'a, C, A>

fn search(&self, query: &str) -> PeopleSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Search all public profiles.

Arguments

query - Specify a query string for full text search of public text in all profiles.

fn list_by_activity(&self, activity_id: &str, collection: &str) -> PeopleListByActivityCall<'a, C, A>

Create a builder to help you perform the following task:

List all of the people in the specified collection for a particular activity.

Arguments

activityId - The ID of the activity to get the list of people for.

collection - The collection of people to list.

fn list(&self, user_id: &str, collection: &str) -> PeopleListCall<'a, C, A>

Create a builder to help you perform the following task:

List all of the people in the specified collection.

Arguments

userId - Get the collection of people for the person identified. Use "me" to indicate the authenticated user.

collection - The collection of people to list.

fn get(&self, user_id: &str) -> PeopleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a person's profile. If your app uses scope https://www.googleapis.com/auth/plus.login, this method is guaranteed to return ageRange and language.

Arguments

userId - The ID of the person to get the profile for. The special value "me" can be used to indicate the authenticated user.