Struct google-plus1::PeopleMethodsBuilder

pub struct PeopleMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on people resources. It is not used directly, but through the Plus hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-plus1" as plus1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use plus1::Plus;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Plus::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `get(...)`, `list(...)`, `list_by_activity(...)` and `search(...)`
// to build up your call.
let rb = hub.people();

Methods

impl<'a, C, NC, A> PeopleMethodsBuilder<'a, C, NC, A>

fn search(&self, query: &str) -> PeopleSearchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Search all public profiles.

fn list_by_activity(&self, activity_id: &str, collection: &str) -> PeopleListByActivityMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all of the people in the specified collection for a particular activity.

fn list(&self, user_id: &str, collection: &str) -> PeopleListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all of the people in the specified collection.

fn get(&self, user_id: &str) -> PeopleGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get a person's profile. If your app uses scope https://www.googleapis.com/auth/plus.login, this method is guaranteed to return ageRange and language.