Struct google_plus1::PersonAgeRange

pub struct PersonAgeRange {
    pub max: Option<i32>,
    pub min: Option<i32>,
}

The age range of the person. Valid ranges are 17 or younger, 18 to 20, and 21 or older. Age is determined from the user's birthday using Western age reckoning.

This type is not used in any activity, and only used as part of another schema.

Fields

max

The age range's upper bound, if any. Possible values include, but are not limited to, the following: - "17" - for age 17 - "20" - for age 20

min

The age range's lower bound, if any. Possible values include, but are not limited to, the following: - "21" - for age 21 - "18" - for age 18